Enemy Controller
Added general purpose enemy controller that can be used on multiple diverse enemy types.
Enemy can patrol  on ground, on platforms and detect changes if moved from one platform to the next.
The enemy can detect the player and move towards the player.
Missing: attack the player when in range.