Updated Dialogue System
Corrected comment errors.

Added a text effect that puts the letters one my one into the dialogue box.
Made it that the cursor is unlocked during dialogue and locked as soon as the dialogue ends.

Factored in code to use multiple names in conversations.

Made the EndDialogue function public to access in a different class.